refactor(setup): setup environment:
- change mocha compiler
- setup environment for development, testing, and production

[Starts #159221700]
