+++ .babelrc
@@ -1,6 +1,6 @@
 {
     "presets": [
-        "es2015"
+        "env"
     ],
     "plugins": ["transform-object-rest-spread"]
 }
