+++ .sequelizerc
@@ -1,8 +1,14 @@
 const path = require('path');
 
+
+let src = 'src'
+
+if(process.env.NODE_ENV === 'production'){
+  src = 'build'
+}
 module.exports = {
-  "config": path.resolve('./build/config', 'config.js'),
-  "models-path": path.resolve('./build/models'),
-  "seeders-path": path.resolve('./build/seeders'),
-  "migrations-path": path.resolve('./build/migrations')
+  "config": path.resolve(`./${src}/config`, 'config.js'),
+  "models-path": path.resolve(`./${src}/models`),
+  "seeders-path": path.resolve(`./${src}/seeders`),
+  "migrations-path": path.resolve(`./${src}/migrations`)
 };
