+++ package.json
@@ -9,24 +9,26 @@
     "yarn": "1.5.1"
   },
   "scripts": {
-    "build": "npm run remove-build && npm run build:server && npm run db:migrate",
+    
+    "build": "npm run remove-build && npm run build:server && npm run db:reset && npm run db:migrate",
     "build:server": "./node_modules/.bin/babel -d ./build ./src -s --copy-files",
     "build-and-start": "npm run build && npm start",
-    "build-and-start:dev": "npm run build && npm run start:dev",
     "codecov": "nyc report --reporter=text-lcov > coverage.lcov && codecov",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "db:migrate": "./node_modules/.bin/sequelize db:migrate",
     "db:migrate:test": "NODE_ENV=test ./node_modules/.bin/sequelize db:migrate",
     "db:reset": "./node_modules/.bin/sequelize db:migrate:undo:all && ./node_modules/.bin/sequelize db:migrate",
     "db:reset:test": "NODE_ENV=test ./node_modules/.bin/sequelize db:migrate:undo:all && NODE_ENV=test ./node_modules/.bin/sequelize db:migrate",
+    "db:seed:database": "NODE_ENV=test ./node_modules/.bin/sequelize db:seed:undo:all && NODE_ENV=test ./node_modules/.bin/sequelize db:seed:all",
     "eslint": "./node_modules/.bin/eslint src",
     "eslint-fix": "./node_modules/.bin/eslint --fix src",
     "heroku-postbuild": "npm run build",
     "remove-build": "./node_modules/.bin/rimraf ./build",
     "start": "NODE_ENV=production ./node_modules/.bin/forever -m 5 ./build/server.js",
-    "start:dev": "NODE_ENV=development nodemon ./build/server",
-    "test": "NODE_ENV=development npm run build && NODE_ENV=test nyc --reporter=html --reporter=text ./node_modules/mocha/bin/_mocha 'test/**/*.js' --compilers js:babel-core/register --timeout 10000 --exit",
-    "test:dev": "NODE_ENV=development npm run build && NODE_ENV=test ./node_modules/mocha/bin/_mocha 'test/**/*.js' --compilers js:babel-core/register --timeout 10000 --exit"
+    "babel-node": "babel-node ./src/server --presets=env",
+    "start:dev": "NODE_ENV=development nodemon --exec npm run babel-node ./src/server",
+    "coverage": "NODE_ENV=test nyc --reporter=html --reporter=text ./node_modules/mocha/bin/_mocha 'test/**/*.js' --compilers js:babel-core/register --timeout 10000 --exit",
+    "test": "npm run db:reset:test && NODE_ENV=test nyc --reporter=html --reporter=text mocha test/**/**.js --require babel-polyfill --require babel-register --timeout 10000 --exit"
   },
   "pre-commit": [
     "eslint"
@@ -41,6 +43,7 @@
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "bcrypt": "^1.0.3",
     "body-parser": "^1.18.2",
@@ -66,6 +69,7 @@
     "webpack": "^4.1.1"
   },
   "devDependencies": {
+    "babel-preset-env": "^1.7.0",
     "codecov": "^3.0.0",
     "coveralls": "^3.0.0",
     "eslint": "^4.9.0",
