+++ src/server.js
@@ -22,6 +22,7 @@ import logger from 'morgan';
 import path from 'path';
 import swaggerJSDoc from 'swagger-jsdoc';
 
+
 import middleware from './middleware';
 import routes from './routes';
